Improve error handling + documentation
