--- emf-compare-adaptor/model_comparison.launch
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<launchConfiguration type="org.eclipse.jdt.launching.remoteJavaApplication">
-    <listAttribute key="org.eclipse.debug.core.MAPPED_RESOURCE_PATHS">
-        <listEntry value="/model_comparison"/>
-    </listAttribute>
-    <listAttribute key="org.eclipse.debug.core.MAPPED_RESOURCE_TYPES">
-        <listEntry value="4"/>
-    </listAttribute>
-    <booleanAttribute key="org.eclipse.jdt.launching.ALLOW_TERMINATE" value="false"/>
-    <mapAttribute key="org.eclipse.jdt.launching.CONNECT_MAP">
-        <mapEntry key="hostname" value="localhost"/>
-        <mapEntry key="port" value="1044"/>
-    </mapAttribute>
-    <stringAttribute key="org.eclipse.jdt.launching.PROJECT_ATTR" value="model_comparison"/>
-    <stringAttribute key="org.eclipse.jdt.launching.VM_CONNECTOR_ID" value="org.eclipse.jdt.launching.socketAttachConnector"/>
-</launchConfiguration>
