+++ emf-compare-adaptor/pom.xml
@@ -1,25 +1,20 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>org.eclipse.model.comparison</groupId>
-  <artifactId>model_comparison</artifactId>
+  <groupId>com.eclipsesource.model.comparison</groupId>
+  <artifactId>model-comparison</artifactId>
   <version>0.0.1-SNAPSHOT</version>
-  <name>model_comparison</name>
+  <name>model-comparison</name>
   <description>Framework for model comparison based on EMF compare</description>
   
   <properties>
-    <maven.compiler.source>1.8</maven.compiler.source>
-    <maven.compiler.target>1.8</maven.compiler.target>
+    <maven.compiler.source>1.11</maven.compiler.source>
+    <maven.compiler.target>1.11</maven.compiler.target>
   </properties>
 
 <repositories>   
 
 </repositories>   
-
-  
-  
   <dependencies>	
-  		<!--  https://github.com/akane34/emf_compare_programatically/blob/master/pom.xml -->
-  		<!-- ************************** EMF COMPARE DEPENDENCIES ********************************* -->	
 		<dependency>
 		    <groupId>com.google.guava</groupId>
 		    <artifactId>guava</artifactId>
@@ -30,7 +25,6 @@
 		    <artifactId>org.eclipse.emf.common</artifactId>
 		    <version>2.19.0</version> <!-- 2.11.1 -->
 		</dependency>
-
 		<dependency>
 		    <groupId>org.eclipse.emf</groupId>
 		    <artifactId>org.eclipse.emf.ecore.xmi</artifactId>
@@ -40,18 +34,12 @@
 		    <groupId>org.eclipse.emf</groupId>
 		    <artifactId>org.eclipse.emf.ecore</artifactId>
 		    <version>2.21.0</version>
-		</dependency>
-		
-		<!--  org.eclipse.ocl 3.5.0 missing -->
-		<!--  org.eclipse.ocl.ecore 3.5.0 missing -->
-		<!--  org.apache.log4j 1.2.15 missing -->
-		
+		</dependency>	
 		<dependency>
 		    <groupId>log4j</groupId>
 		    <artifactId>log4j</artifactId>
 		    <version>1.2.17</version>
 		</dependency>
-	    
 		<dependency>
 		    <groupId>org.eclipse.emf</groupId>
 		    <artifactId>compare</artifactId>
@@ -59,7 +47,6 @@
 		    <scope>system</scope>
 		    <systemPath>${project.basedir}/lib/org.eclipse.emf.compare_3.5.3.201911051515.jar</systemPath>
 		</dependency>
-		
 		<dependency>
 		    <groupId>org.eclipse.emf</groupId>
 		    <artifactId>compare.rcp.ui</artifactId>
@@ -67,7 +54,6 @@
 		    <scope>system</scope>
 		    <systemPath>${project.basedir}/lib/org.eclipse.emf.compare.rcp.ui_4.4.2.201911051515.jar</systemPath>
 		</dependency>
-		
 		<dependency>
 		    <groupId>org.eclipse.emf</groupId>
 		    <artifactId>compare.ide.ui</artifactId>
@@ -75,7 +61,6 @@
 		    <scope>system</scope>
 		    <systemPath>${project.basedir}/lib/org.eclipse.emf.compare.ide.ui_4.4.3.201909101346.jar</systemPath>
 		</dependency>
-		
 		<dependency>
 		    <groupId>org.eclipse.compare</groupId>
 		    <artifactId>org.eclipse.compare</artifactId>
@@ -83,20 +68,16 @@
 		    <scope>system</scope>
 		    <systemPath>${project.basedir}/lib/org.eclipse.compare_3.7.1000.v20200511-1203.jar</systemPath>
 		</dependency>
-		
-		
 		<dependency>
 		    <groupId>org.eclipse.emf</groupId>
 		    <artifactId>org.eclipse.emf.edit</artifactId>
 		    <version>2.16.0</version>
 		</dependency>
-		
 		<dependency>
 		    <groupId>org.eclipse.core</groupId>
 		    <artifactId>org.eclipse.core.runtime</artifactId>
 		    <version>3.7.0</version>
 		</dependency>
-		
 		<dependency>
 		    <groupId>org.eclipse.emf</groupId>
 		    <artifactId>compare.edit</artifactId>
@@ -104,7 +85,6 @@
 		    <scope>system</scope>
 		    <systemPath>${project.basedir}/lib/org.eclipse.emf.compare.edit_4.3.1.201909101346.jar</systemPath>
 		</dependency>
-		
 		<dependency>
 		    <groupId>org.eclipse.emf</groupId>
 		    <artifactId>compare.ide</artifactId>
@@ -112,26 +92,15 @@
 		    <scope>system</scope>
 		    <systemPath>${project.basedir}/lib/org.eclipse.emf.compare.ide_3.4.3.202008311302.jar</systemPath>
 		</dependency>
-		
 		<dependency>
 		    <groupId>org.eclipse.core</groupId>
 		    <artifactId>org.eclipse.core.resources</artifactId>
 		    <version>3.7.100</version>
 		</dependency>
-		
-		
-		
-		<!-- for JSON conversion ??? -->
 		<dependency>
 		    <groupId>com.google.code.gson</groupId>
 		    <artifactId>gson</artifactId>
 		    <version>2.8.6</version>
-		</dependency>
-		
-		
-				
+		</dependency>			
 	</dependencies> 
-    
-
-
 </project>
