+++ emf-compare-adaptor/src/main/java/com/eclipsesource/model/comparison/AttributeChangeDiffObject.java
@@ -1,0 +1,39 @@
+/********************************************************************************
+ * Copyright (c) 2021 EclipseSource and others.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License v. 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0.
+ *
+ * This Source Code may also be made available under the following Secondary
+ * Licenses when the conditions for such availability set forth in the Eclipse
+ * Public License v. 2.0 are satisfied: GNU General Public License, version 2
+ * with the GNU Classpath Exception which is available at
+ * https://www.gnu.org/software/classpath/license.html.
+ *
+ * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
+ ********************************************************************************/
+package com.eclipsesource.model.comparison;
+
+import org.eclipse.emf.ecore.EObject;
+
+public class AttributeChangeDiffObject extends DiffObject {
+	private String value;
+
+	public AttributeChangeDiffObject(EObject object, String value, String uuid) {
+		super();
+		this.object = object;
+		this.value = value;
+		this.uuid = uuid;
+		this.graphicalId = UUID_Provider.getGraphicalUUID(object);
+	}
+
+	public String getValue() {
+		return value;
+	}
+
+	public void setValue(String value) {
+		this.value = value;
+	}
+
+}
