+++ emf-compare-adaptor/src/main/java/com/eclipsesource/model/comparison/ObjectManager.java
@@ -1,0 +1,55 @@
+/********************************************************************************
+ * Copyright (c) 2021 EclipseSource and others.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License v. 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0.
+ *
+ * This Source Code may also be made available under the following Secondary
+ * Licenses when the conditions for such availability set forth in the Eclipse
+ * Public License v. 2.0 are satisfied: GNU General Public License, version 2
+ * with the GNU Classpath Exception which is available at
+ * https://www.gnu.org/software/classpath/license.html.
+ *
+ * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
+ ********************************************************************************/
+package com.eclipsesource.model.comparison;
+
+public abstract class ObjectManager<T extends DiffObject> {
+	protected String uuid;
+	protected T source;
+	protected T target;
+	protected T base;
+
+	public String getUuid() {
+		return uuid;
+	}
+
+	public void setUuid(String uuid) {
+		this.uuid = uuid;
+	}
+
+	public T getSource() {
+		return source;
+	}
+
+	public void setSource(T source) {
+		this.source = source;
+	}
+
+	public T getTarget() {
+		return target;
+	}
+
+	public void setTarget(T target) {
+		this.target = target;
+	}
+
+	public T getBase() {
+		return base;
+	}
+
+	public void setBase(T base) {
+		this.base = base;
+	}
+}
